Extendenable tranceiver
