+++ spinnman/extended/version3transceiver.py
@@ -1,0 +1,19 @@
+# Copyright (c) 2023 The University of Manchester
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from .extended_transceiver import ExtendedTransceiver
+from spinnman.transceiver.version3transceiver import Version3Transceiver
+
+
+class ExtendedVersion3Transceiver(Version3Transceiver, ExtendedTransceiver):
+    pass
