+++ spinnman/transceiver/version3transceiver.py
@@ -17,18 +17,19 @@ from spinnman.data import SpiNNManDataView
 from spinnman.transceiver.base_transceiver import BaseTransceiver
 
 
-class Version5Transceiver(BaseTransceiver):
+class Version3Transceiver(BaseTransceiver):
     """
-    Implementation of the Transceiver classes for Version 5 boards
+    Implementation of the Transceiver classes for Version 3 boards
+
+    This class should ONLY be created via by transceiver_factory.py
     """
-    pass
 
     @overrides(BaseTransceiver.__init__)
     def __init__(self, connections=None):
         super().__init__(connections)
-        assert SpiNNManDataView.get_machine_version().number == 5
+        assert SpiNNManDataView.get_machine_version().number == 3
 
     @property
     @overrides(BaseTransceiver.boot_led_0_value)
     def boot_led_0_value(self):
-        return 0x00000001
+        return 0x00000502
